polymer-resin: require an explicit install
This makes it easier to run in an experiment framework.

I had hoped initially that I could simply provide an HTML importable module.

That doesn't work because of idiosyncracies of HTML import.
The solution to that is an empty script tag.
That empty script tag is less likely to be written out if there's something
in it, like a call to install.

This bundles several changes:
1. Adds security.polymer_resin.install
2. Reworks existing public API to work via a configuration object passed
   to insall.
3. Modifies existing tests to install.
4. Flesh out docs with install and the configuration.
Tested:
Reran project tests.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157182380