Sync package descriptions
Changed bower.json to use the same web-component-tester version as package.json.

Changed package.json to export the same file list as bower.json.

The package.json has to be overly specific because there is no test/
directory as is typical in NPM projects due to Polymer BUILD rules
requiring less tweaking when tests are in the same directory as the
component they test.

For background, bower (deprecated) is a package system used by Gerrit (still on
Polymer 1) and npm is now used by Polymer 3.

PiperOrigin-RevId: 207302311